Struct goatns::config::ConfigFile

pub struct ConfigFile {

    pub hostname: String,
    pub address: String,
    pub port: u16,
    pub capture_packets: bool,
    pub log_level: String,
    pub tcp_client_timeout: u64,
    pub enable_hinfo: bool,
    pub sqlite_path: String,
    pub zone_file: Option<String>,
    pub allowed_tlds: Vec<String>,
    pub ip_allow_lists: IPAllowList,
    pub enable_api: bool,
    pub api_port: u16,
    pub api_tls_cert: PathBuf,
    pub api_tls_key: PathBuf,
    pub api_static_dir: String,
    api_cookie_secret: String,
    pub oauth2_client_id: String,
    pub oauth2_redirect_url: Url,
    pub oauth2_secret: String,
    pub oauth2_config_url: String,
    pub oauth2_user_scopes: Vec<String>,
    pub sql_log_statements: bool,
    pub sql_log_slow_duration: u64,
    pub sql_db_cleanup_seconds: u64,
    pub admin_contact: ContactDetails,
    pub user_auto_provisioning: bool,

}

The main config blob for GoatNS, write this as a JSON file and load it and it’ll make things go.

Fields
hostname: String

The server’s hostname when generating an SOA record, defaults to the results of gethostname()

address: String

DNS listener address, default is 127.0.0.1

port: u16

Listen for DNS queries on this port, default is 15353

capture_packets: bool

If we should capture packets on request/response

log_level: String

Default is “DEBUG”

tcp_client_timeout: u64

How long until we drop TCP client connections, defaults to 5 seconds.

enable_hinfo: bool

Enable a HINFO record at hinfo.goat

sqlite_path: String

The location for the zone sqlite file

zone_file: Option<String>

Where the JSON zone file is

allowed_tlds: Vec<String>

List of “valid” TLDs - if this is empty let anything be created

ip_allow_lists: IPAllowList

IP Allow lists

enable_api: bool

Do you really want an API?

api_port: u16

API / Web UI Port

api_tls_cert: PathBuf

Certificate path

api_tls_key: PathBuf

TLS key path

api_static_dir: String

Static File Directory for api things

api_cookie_secret: String

Secret for cookie storage - it’ll randomly generate on startup by default

oauth2_client_id: String

OAuth2 Resource server name

oauth2_redirect_url: Url

If your instance is behind a proxy/load balancer/whatever, you need to specify this, eg https://example.com:12345

oauth2_secret: String

Oauth2 Secret

oauth2_config_url: String

OIDC Discovery URL, eg for Kanidm you’d use https://idm.example.com/oauth2/openid/:client_id:/.well-known/openid-configuration

oauth2_user_scopes: Vec<String>

A list of scopes to request from the IdP

sql_log_statements: bool

Log things sometimes

sql_log_slow_duration: u64

When queries take more than this many seconds, log them

sql_db_cleanup_seconds: u64

Clean up sessions table every n seconds

admin_contact: ContactDetails

Administrator contact details

user_auto_provisioning: bool

Allow auto-provisioning of users

Implementations

impl ConfigFile

pub fn as_json_pretty(&self) -> String

JSONify the configfile in a pretty way using serde

pub fn dns_listener_address(&self) -> Result<SocketAddr, Option<String>>

Get a bindable SocketAddr for use in the DNS listeners

pub fn api_listener_address(&self) -> SocketAddr

get a string version of the listener address

pub fn api_cookie_secret(&self) -> &[u8]

It’s a sekret!

pub fn status_url(&self) -> Url

pub async fn get_tls_config(&self) -> RustlsConfig

pub async fn check_config(
config: CowCellWriteTxn<'_, ConfigFile>
) -> Result<(), Vec<String>>

pub fn try_as_cowcell(
config_path: Option<&String>
) -> Result<CowCell<ConfigFile>, Error>

Uses Self::try_from and wraps it in a CowCell (moo)

The default locations are ~/.config/goatns.json and ./goatns.json.

pub fn try_from(config_path: Option<&String>) -> Result<ConfigFile, Error>

Loads the configuration from a given file or from some default locations.

The default locations are ~/.config/goatns.json and ./goatns.json.