Module goatns::enums

Enums

Agent

AgentState

ContactDetails

ContactDetailsDeserializerError

Errors returned from trying to turn a String into a ContactDetails

OpCode

A four bit field that specifies kind of query in this message. This value is set by the originator of a query and copied into the response.

PacketType

Rcode

Response code, things like NOERROR, FORMATERROR, SERVFAIL etc.

RecordClass

CLASS fields appear in resource records, most entries should be IN, but CHAOS is typically used for management-layer things. Ref RFC1035 3.2.4.

RecordType

RRType, eg A, NS, MX, etc

SystemState

Constants

OP_CODE_ALL 🔒

PACKET_TYPE_ALL 🔒

RCODE_ALL 🔒