Enum goatns::resourcerecord::InternalResourceRecord

pub enum InternalResourceRecord {

    A {
        address: u32,
        ttl: u32,
        rclass: RecordClass,
    },
    AAAA {
        address: u128,
        ttl: u32,
        rclass: RecordClass,
    },
    AXFR {
        ttl: u32,
        rclass: RecordClass,
    },
    CAA {
        flag: u8,
        tag: DNSCharString,
        value: Vec<u8>,
        ttl: u32,
        rclass: RecordClass,
    },
    CNAME {
        cname: DomainName,
        ttl: u32,
        rclass: RecordClass,
    },
    LOC {
        ttl: u32,
        rclass: RecordClass,
        version: u8,
        size: u8,
        horiz_pre: u8,
        vert_pre: u8,
        latitude: u32,
        longitude: u32,
        altitude: i32,
    },
    NAPTR {
        ttl: u32,
        rclass: RecordClass,
        domain: DomainName,
        order: u16,
        preference: u16,
        flags: String,
    },
    NS {
        nsdname: DomainName,
        ttl: u32,
        rclass: RecordClass,
    },
    SOA {
        zone: DomainName,
        mname: DomainName,
        rname: DomainName,
        serial: u32,
        refresh: u32,
        retry: u32,
        expire: u32,
        minimum: u32,
        rclass: RecordClass,
    },
    PTR {
        ptrdname: DomainName,
        ttl: u32,
        rclass: RecordClass,
    },
    HINFO {
        cpu: Option<DNSCharString>,
        os: Option<DNSCharString>,
        ttl: u32,
        rclass: RecordClass,
    },
    MX {
        preference: u16,
        exchange: DomainName,
        ttl: u32,
        rclass: RecordClass,
    },
    TXT {
        txtdata: DNSCharString,
        ttl: u32,
        class: RecordClass,
    },
    URI {
        priority: u16,
        weight: u16,
        target: DNSCharString,
        ttl: u32,
        rclass: RecordClass,
    },
    InvalidType,

}

Internal representation of a resource record

Variants

A

Fields

address: u32

ttl: u32

rclass: RecordClass

A single host address

AAAA

Fields

address: u128

ttl: u32

rclass: RecordClass

AXFR

Fields

ttl: u32

rclass: RecordClass

CAA

Fields

flag: u8

tag: DNSCharString

Tags MAY contain ASCII characters “a” through “z”, “A” through “Z”, and the numbers 0 through 9. Tags MUST NOT contain any other characters. Matching of tags is case insensitive.

value: Vec<u8>

A sequence of octets representing the Property Value. Property Values are encoded as binary values and MAY employ sub‑formats.

ttl: u32

rclass: RecordClass

CNAME

Fields

cname: DomainName

ttl: u32

rclass: RecordClass

LOC

Fields

ttl: u32

rclass: RecordClass

version: u8

Version number of the representation. This must be zero. Implementations are required to check this field and make no assumptions about the format of unrecognized versions.

size: u8

horiz_pre: u8

vert_pre: u8

latitude: u32

longitude: u32

altitude: i32

NAPTR

Fields

ttl: u32

rclass: RecordClass

domain: DomainName

Domain - The domain name to which this resource record refers. This is the ‘key’ for this entry in the rule database. This value will either be the first well known key (<something>.uri.arpa for example) or a new key that is the output of a replacement or regexp rewrite. Beyond this, it has the standard DNS requirements.

order: u16

preference: u16

flags: String

NS

Fields

nsdname: DomainName

ttl: u32

rclass: RecordClass

SOA

Fields

zone: DomainName

mname: DomainName

The domain-name of the name server that was the original or primary source of data for this zone.

rname: DomainName

A domain-name which specifies the mailbox of the person responsible for this zone. eg: dns.example.com is actually dns@example.com

serial: u32

refresh: u32

retry: u32

expire: u32

minimum: u32

rclass: RecordClass

PTR

Fields

ptrdname: DomainName

ttl: u32

rclass: RecordClass

HINFO

Fields

cpu: Option<DNSCharString>

os: Option<DNSCharString>

ttl: u32

rclass: RecordClass

RFC1035

MX

Fields

preference: u16

exchange: DomainName

ttl: u32

rclass: RecordClass

TXT

Fields

txtdata: DNSCharString

ttl: u32

class: RecordClass

URI

Fields

priority: u16

weight: u16

target: DNSCharString

ttl: u32

rclass: RecordClass

InvalidType

Implementations

impl InternalResourceRecord

pub fn as_bytes(self: &InternalResourceRecord, question: &Vec<u8>) -> Vec<u8>

pub fn hexdump(self)

pub fn ttl(&self) -> &u32