Module goatns::utils

Functions

clap_parser

cli_commands

dms_to_u32

turn a degrees/minutes/seconds format into unsigned 32-bit integer matching the format required for a DNS LOC record

export_zone_file

dump a zone to a file

find_tail_match

If you have a name and a target and want to see if you can find a chunk of the target that the name ends with, this is your function!

hexdump

dumps the bytes out as if you were using some kind of fancy packet-dumper

import_zones

loc_size_to_u8

converts size/precision X * 10**Y(cm) to 0xXY This code is ported from the C code in RFC1876 (Appendix A, precsize_aton)

name_as_bytes

reply_builder

Want a generic empty reply with an ID and an RCODE? Here’s your function.

reply_nxdomain

seven_dot_three_conversion 🔒

does the conversion from “example.com” to “7example3com” BUT DOES NOT DO THE TRAILING NULL BECAUSE REASONS

vec_find