Crate goatns

Modules

api

Configuration and management API

config

Configuration handling for the server

datastore

The data-storing backend for zone information and (eventually) caching.

db

enums

packet_dumper

reply

resourcerecord

serializers

servers

tests 🔒

utils

zones

Structs

Header

The header of a DNS transmission, either a Query or Reply. Ref RFC1035 section 4.1.1.

Question

A DNS Question section, from Ref RFC1035 section 4.1.2 “Question section format”.

ResourceRecord

The answer, authority, and additional sections all share the same format: a variable number of resource records, where the number of records is specified in the corresponding count field in the header.

Constants

HEADER_BYTES

The size of a DNS request header

MAX_IN_FLIGHT

Internal limit of in-flight requests

REPLY_TIMEOUT_MS

The default “cancel a server response” timeout

UDP_BUFFER_SIZE

The maximum size of a UDP packet https://dnsflagday.net/2020/#dns-flag-day-2020

Functions

get_question_qname

Returns a Vec representation of the name (ie, example.com = [101, 120, 97, 109, 112, 108, 101, 46, 99, 111, 109])