Module goatns::resourcerecord

Structs

CAA_TAG_VALIDATOR 🔒

DNSCharString

<character-string> is a single length octet followed by that number of characters. <character-string> is treated as binary information, and can be up to 256 characters in length (including the length octet).

DomainName

FileLocRecord

This represents a LOC record in a zone file

LOC_REGEX 🔒

LocRecord

URI_RECORD 🔒

Enums

InternalResourceRecord

Internal representation of a resource record

Constants

DEFAULT_LOC_HORIZ_PRE 🔒

DEFAULT_LOC_SIZE 🔒

DEFAULT_LOC_VERT_PRE 🔒

Traits

SetTTL

Functions

check_long_labels

tests to ensure that no label in the name is longer than 63 octets (bytes)