Module goatns::resourcerecord

Structs

CAA_TAG_VALIDATOR ğŸ”’

DNSCharString

<character-string> is a single length octet followed by that number of characters. <character-string> is treated as binary information, and can be up to 256 characters in length (including the length octet).

DomainName

FileLocRecord

This represents a LOC record in a zone file

LOC_REGEX ğŸ”’

LocRecord

URI_RECORD ğŸ”’

Enums

InternalResourceRecord

Internal representation of a resource record

Constants

DEFAULT_LOC_HORIZ_PRE ğŸ”’

DEFAULT_LOC_SIZE ğŸ”’

DEFAULT_LOC_VERT_PRE ğŸ”’

Statics

LAZY ğŸ”’

LAZY ğŸ”’

LAZY ğŸ”’

Traits

SetTTL

Functions

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

check_long_labels

tests to ensure that no label in the name is longer than 63 octets (bytes)